MWPW-134968 - [Blog2Milo] Broken links on Homepage
- address PR feedback
- move the utility function to its own helpers.js below /libs/utils
- rename utility function
- error handling for new URL()
