+++ libs/blocks/recommended-articles/recommended-articles.js
@@ -1,7 +1,7 @@
 import { createTag, getMetadata, getConfig } from '../../utils/utils.js';
 import fetchTaxonomy from '../../scripts/taxonomy.js';
 import { replaceKey } from '../../features/placeholders.js';
-import { updateLinkWithConfiguredLangRoot } from '../utils/utils.js';
+import { updateLinkWithLangRoot } from '../../utils/helpers.js';
 
 async function getArticleDetails(article) {
   const path = new URL(article.href).pathname;
@@ -42,7 +42,7 @@ function getDecoratedCards(articles, taxonomy) {
     const cardImage = createTag('div', { class: 'article-card-image' }, imageEl);
     const cardBody = createTag('div', { class: 'article-card-body' });
     const categoryTaxonomy = taxonomy.get(category) || 'News';
-    const categoryLink = createTag('a', { href: updateLinkWithConfiguredLangRoot(categoryTaxonomy.link) }, categoryTaxonomy.name);
+    const categoryLink = createTag('a', { href: updateLinkWithLangRoot(categoryTaxonomy.link) }, categoryTaxonomy.name);
     const categoryEl = createTag('p', { class: 'article-card-category' }, categoryLink);
     const titleEl = createTag('h3', null, title);
     const descriptionEl = createTag('p', { class: 'article-card-description' }, description);
