+++ libs/blocks/utils/utils.js
@@ -1,5 +1,3 @@
-import { getMetadata } from '../../utils/utils.js';
-
 const getLocalStorage = (key) => {
   const data = window.localStorage.getItem(key);
   try {
@@ -28,12 +26,4 @@ const fetchWithTimeout = async (resource, options = {}) => {
   return response;
 };
 
-const updateLinkWithConfiguredLangRoot = (link) => {
-  const langRoot = getMetadata('lang-root');
-  if (!langRoot) return link;
-  const url = new URL(link);
-  url.pathname = `${langRoot}${url.pathname}`;
-  return url.href;
-};
-
-export { getLocalStorage, setLocalStorage, fetchWithTimeout, updateLinkWithConfiguredLangRoot };
+export { getLocalStorage, setLocalStorage, fetchWithTimeout };
